requirements.txt

streamlit
pandas
numpy
matplotlib
requests
pyotp
smartapi-python
